pyproject.toml
[tool.poetry]
name = "lovinators-ultimate-sitemap-parser"
version = "0.1.0"
description = ""
authors = [
    "Joakim Hellsén <tlovinator@gmail.com>",
    "Linas Valiukas <linas@media.mit.edu>",
    "Hal Roberts <hroberts@cyber.law.harvard.edu>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1.11.0"
pytest = "^7.4.3"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true
preview = true
select = ["ALL"]
ignore = ["D100", "D104", "CPY001", "ERA001", "TD002", "FIX002", "TD003"]
unsafe-fixes = true

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests...
    "ARG",  # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT",  # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
]
